Load test of zkevm-node with sequencer
- Move `contract-bindings` to `zkevm-contract-bindings`.

I encountered errors with the docker image from within the static shell
because Cargo was trying to clone dependencies and this doesn't work
from within the cross compile dev shell. I thought it could be the crate
rename in Cargo.toml (to avoid having two crates names
`contract-bindings`) that could cause this so went on to remove that
naming conflict. In the end, this issue was resolved by keeping the
`HOME` env var and setting `CARGO_HOME` in the cross-shell.

For now I kept the crate rename because I think it's a bit less
confusing this way.

- Start up the full demo with docker compose from within rust.
- Save a test plan (currently only includes groups of 0 to 10
  Eth transfers).
- Load test plan for replay.
- Basic CLI to run and replay the load test.