Consolidate HotShot address environment variables
We had two environment variables:
* ESPRESSO_SEQUENCER_HOTSHOT_ADDRESS
* ESPRESSO_ZKEVM_HOTSHOT_ADDRESS
which always must be set to the same value. This is fine when we
set one to the value of another in .env, but it means that when
overriding one of the values in the environment, we must always
remember to override the other as well. Indeed, the demo was not
working for me because `polygon_zkevm.rs` was setting
`ESPRESSO_ZKEVM_HOTSHOT_ADDRESS` to the deployed address from
`l1.hotshot.address()`, but failing to set
`ESPRESO_SEQUENCER_HOTSHOT_ADDRESS`.

To fix the demo and remove this footgun, I removed
`ESPRESSO_ZKEVM_HOTSHOT_ADDRESS` and changed the code to use
`ESPRESSO_SEQUENCER_HOTSHOT_ADDRESS` everywhere.