Adjust demo to demonstrate preconfirmations
Increase L1 block time and decrease HotShot block time so that
preconfirmations are more noticeable. The L1 block time of 5
seconds is still significantly lower than the Ethereum block time
(12 seconds) and the Ethereum finality latency (12 minutes), so
we are not cheating.

Both block times are overridable via the environment.

For some reason the geth node ignores the `--dev.period` argument
when started via our docker-compose setup, even though directly
invoking geth with the argument does adjust the block time. I
suspect this has to do with the saved state in the mock L1 Docker
image, which may be encoding a different block time somehow. I will
continue to investigate this. In the mean time the demo can be run
with `just demo --l1-backend anvil`.