Add end-to-end test for preconfirmations
The test sets up an L2 system with a regular node and a
preconfirmations node. It sequences a transaction, checks that
both nodes finalize the transaction, and asserts that the
preconfirmation node finalized it first.

With a HotShot block time of 1s and L1 block time of 5s, this
should be highly deterministic. Indeed, on my machine, the
preconfirmation arrives 5-10s before the final confirmation.