Separate related services into modular profiles
The demo consists basically of 5 groups of related services, which
are mostly independent of each other:
* The base services (sequencer and L1)
* zkevm1 node and dependencies
* zkevm1 node with preconfirmations and dependencies
* zkevm2 node and dependencies
* zkevm2 node with preconfirmations and dependencies

This changes uses docker-compose's profiles feature to explicitly
group each of these modules (except the base services, which belong
to the implicit default protocol which is always started automatically).
This makes it easy to run just one or any combination of the components,
without having to manually list out the services you want to start.

Since the demo now includes 4 provers and uses quite a bit of memory,
this will help users on lower-end machines run a lighter weight
version of the demo.