Put the Polygon block timestamp back the way it was.
Currently the sequencer emits a genesis block with timestamp 0.
The zkevm-node does not properly handle blocks with timestamp before
its own genesis (it should ignore them, but doesn't). Both of these
things should probably be changed, but for now as a workaround we
use the timestamp assigned by HotShot, not the sequencer, which is
never 0.