Fix end-to-end tests
The zkevm-node requires accurate L1 block numbers from HotShot to
function properly. This requires each sequencer node to be
configured with the L1 provider URL. Unfortunately, due to
limitations in the current HotShot API (https://github.com/EspressoSystems/HotShot/issues/1538)
the only way to pass this configuration into the consensus hooks
is via a process-global variable set based on the environment.

This is obviously not conducive to running multiple tests in the
same process, as we currently do. This change reworks the end-to-end
tests (the ones that run the zkevm-node) to run the sequencer and
zkevm adaptor in Docker as well, rather than on the host, so that
we can more easily configure the sequencers for each test separately.

I also made several improvements to our docker-compose automation
to set the L1 block time (which varies depending on the test) and
to avoid tearing down the L1 when starting up the L2 services. The
changes require docker-compose 2.4 or greater. This should be fine,
since the latest version is 2.20, which is widely available, even
on Mac. I did have to update my Docker Desktop installation to get
the newer version, though.