Increase timeout for CI tests
The new reorg test takes quite a long time. If this works, I will
reorganize the test suite to run this in a scheduled job or
something. It's too long and not really necessary to run every PR,
since it's not a great regression test anyways (you need to look
at the logs to see if the reorg problem occurred). But I want to
get it passing in CI at least once before scaling it back.