Remove --abort-on-container-exit in end-to-end tests.
We want these commands to continue running throughout the test, so
that we can see logs from the services. However, if we run them
with --abort-on-container-exit, they will abort when we start
shutting down the services normally at the end of the tests. We
will then have 3 processes (L1 services, L2 services, and docker
compose down) all trying to shut down the same services, and some
of them will fail, leading to messy shutdown.